int id -> long id

+++ LaunchServer/source/response/PingResponse.java
@@ -8,7 +8,7 @@ import launcher.serialize.HOutput;
 import launchserver.LaunchServer;
 
 public final class PingResponse extends Response {
-	public PingResponse(LaunchServer server, int id, HInput input, HOutput output) {
+	public PingResponse(LaunchServer server, long id, HInput input, HOutput output) {
 		super(server, id, input, output);
 	}
