+++ LaunchServer/source/response/auth/AuthResponse.java
@@ -18,7 +18,7 @@ import launchserver.response.Response;
 import launchserver.response.profile.ProfileByUUIDResponse;
 
 public final class AuthResponse extends Response {
-	public AuthResponse(LaunchServer server, int id, HInput input, HOutput output) {
+	public AuthResponse(LaunchServer server, long id, HInput input, HOutput output) {
 		super(server, id, input, output);
 	}
