+++ LaunchServer/source/response/profile/ProfileByUsernameResponse.java
@@ -10,7 +10,7 @@ import launchserver.LaunchServer;
 import launchserver.response.Response;
 
 public final class ProfileByUsernameResponse extends Response {
-	public ProfileByUsernameResponse(LaunchServer server, int id, HInput input, HOutput output) {
+	public ProfileByUsernameResponse(LaunchServer server, long id, HInput input, HOutput output) {
 		super(server, id, input, output);
 	}
